symbolmanage

+++ Slot/row_test.go
@@ -6,6 +6,6 @@ import (
 )
 
 func TestRow(t *testing.T) {
-	a := NewSymbol("A", NORMAL, []int{0, 0, 10, 15, 20})
+	a := NewSymbol("A", NORMAL, []float64{0, 0, 10, 15, 20})
 	fmt.Println(a)
 }
